api routes structured
